requirements.txt
python3-flask>=0.10.0
python3-sqlalchemy>=1.0.11
python3-flask-restful>=0.3.4e
python3-msgpack>= 0.4.6